feat: move physics component into physics manager
