+++ packages/core/src/ComponentsManager.ts
@@ -2,7 +2,6 @@ import { Vector3 } from "@oasis-engine/math";
 import { Camera } from "./Camera";
 import { Component } from "./Component";
 import { DisorderedArray } from "./DisorderedArray";
-import { Collider, CharacterController } from "./physics";
 import { Renderer } from "./Renderer";
 import { RenderContext } from "./RenderPipeline/RenderContext";
 import { Script } from "./Script";
@@ -33,10 +32,6 @@ export class ComponentsManager {
   // Delay dispose active/inActive Pool
   private _componentsContainerPool: Component[][] = [];
 
-  // Physics
-  private _colliders: DisorderedArray<Collider> = new DisorderedArray();
-  private _characterControllers: DisorderedArray<CharacterController> = new DisorderedArray();
-
   addRenderer(renderer: Renderer) {
     renderer._rendererIndex = this._renderers.length;
     this._renderers.add(renderer);
@@ -57,28 +52,6 @@ export class ComponentsManager {
     const replaced = this._onStartScripts.deleteByIndex(script._onStartIndex);
     replaced && (replaced._onStartIndex = script._onStartIndex);
     script._onStartIndex = -1;
-  }
-
-  addCollider(collider: Collider) {
-    collider._index = this._colliders.length;
-    this._colliders.add(collider);
-  }
-
-  removeCollider(collider: Collider): void {
-    const replaced = this._colliders.deleteByIndex(collider._index);
-    replaced && (replaced._index = collider._index);
-    collider._index = -1;
-  }
-
-  addCharacterController(controller: CharacterController) {
-    controller._index = this._characterControllers.length;
-    this._characterControllers.add(controller);
-  }
-
-  removeCharacterController(controller: CharacterController) {
-    let replaced = this._characterControllers.deleteByIndex(controller._index);
-    replaced && (replaced!._index = controller._index);
-    controller._index = -1;
   }
 
   addOnUpdateScript(script: Script) {
@@ -287,27 +260,6 @@ export class ComponentsManager {
     for (let i = scripts.length - 1; i >= 0; --i) {
       const script = scripts.get(i);
       script._waitHandlingInValid || script.onEndRender(camera);
-    }
-  }
-
-  callColliderOnUpdate(): void {
-    const elements = this._colliders._elements;
-    for (let i = this._colliders.length - 1; i >= 0; --i) {
-      elements[i]._onUpdate();
-    }
-  }
-
-  callColliderOnLateUpdate(): void {
-    const elements = this._colliders._elements;
-    for (let i = this._colliders.length - 1; i >= 0; --i) {
-      elements[i]._onLateUpdate();
-    }
-  }
-
-  callCharacterControllerOnLateUpdate() {
-    let elements = this._characterControllers._elements;
-    for (let i = this._characterControllers.length - 1; i >= 0; --i) {
-      elements[i]!._onLateUpdate();
     }
   }
